requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.7
alembic==1.12.1
celery==5.3.4
redis==5.0.1
python-jose[cryptography]==3.4.0
python-multipart==0.0.19
PyPDF2==3.0.1
python-docx==1.1.0
minio==7.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2